pyproject.toml
[project]
name = "open-lifeworlds-data-product-berlin-points-of-interest"
version = "0.1.0"
description = "Data product combining Berlin point-of-interest data and geodata"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "open-lifeworlds-python-lib",
    "rich>=14.1.0",
    "shapely>=2.1.1",
]

[tool.uv.sources]
open-lifeworlds-python-lib = { git = "https://github.com/open-lifeworlds/open-lifeworlds-python-lib.git" }
